+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [0.5] - 2019-05-21
+### Changed
+- `JSONSerializer`: added serialization of `Pair` and `Triple`
+- `JSONSerializerTest`: corresponding tests
+- `JSONDeserializer`: added deserialization of `Pair` and `Triple`
+- `JSONDeserializerTest`: corresponding tests
+- `JSONDeserializer`: added checking of types marked nullable
+- `JSONDeserializerTest`: corresponding tests
+- `JSONDeserializer`: added use of custom deserialization from `JSONConfig`
+- `JSONDeserializerTest`: corresponding tests
+- `JSONSerializer`: added use of custom serialization from `JSONConfig`
+- `JSONSerializerTest`: corresponding tests
+
+### Added
+- `JSONConfig`: new configuration class
+
 ## [0.4] - 2019-05-14
 ### Changed
 - `JSONDeserializerTest`: modified to use `KType` values